requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dateparser==1.2.2
Flask==3.1.2
hvac==2.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
prometheus-api-client==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-nomad==2.1.0
pytz==2025.2
regex==2025.11.3
requests==2.32.5
six==1.17.0
tzlocal==5.3.1
urllib3==2.6.2
Werkzeug==3.1.4
